exercise 1 corrections
