+++ exercises/matthew_bowles/ex01_basics/inc/Io.h
@@ -1,0 +1,27 @@
+#ifndef IO_H
+#define IO_H
+
+#include "WordManipulation.h"
+#include <fstream>
+#include <iostream>
+#include <iomanip>
+#include <string>
+#include <map>
+#include <vector>
+#include <algorithm>
+
+using namespace std;
+
+//loads word input from a file 
+map<string, int> loadFile(const string& path);
+
+//splits a line of the input into words
+void splitLine(string &line, map<string, int>& words);
+
+//sorts map of words into descending order by value
+vector<pair<string, int>> sortWords(const map<string, int>& words);
+
+//outputs the words to a file
+void outputCount(const vector<pair<string, int>>& words);
+
+#endif
